move best quote
